requirements.txt
python-telegram-bot==20.3
Flask==2.2.5
Pillow==10.0.0
SQLAlchemy==2.0.20
